chore: move modifications out of plugin.rb and introduce improved code structure
